Refactor: 일정공유방 페이지 사이드바 컴포넌트 영역 리팩토링
일정공유방 페이Á 사이드바 컴포넌트 영역 리팩토링
