+++ src/components/planShareRoom/sideBar/placeSearchArea/PlaceSearchArea.tsx
@@ -4,10 +4,10 @@ import PlaceSearchLists from "./PlaceSearchLists";
 
 const PlaceSearchArea = () => {
   return (
-    <div className="absolute left-[25rem] top-0 w-[25rem] min-h-screen border-2">
-      <div className="flex flex-col items-center h-[140px] bg-main-color">
-        <PlaceSearchCategory />
+    <div className="">
+      <div className="flex flex-col items-center h-[140px] bg-[#00AFFF]">
         <PlaceSearchBar />
+        <PlaceSearchCategory />
       </div>
       <PlaceSearchLists />
     </div>
