+++ src/components/planShareRoom/sideBar/placeSearchArea/PlaceSearchBar.tsx
