+++ src/components/planShareRoom/sideBar/placeSearchArea/PlaceSearchCategory.tsx
